+++ src/components/settings_modal/tabs/general_tab.vue
@@ -192,6 +192,16 @@
           </BooleanSetting>
         </li>
         <li>
+          <BooleanSetting path="alwaysShowNewPostButton">
+            {{ $t('settings.always_show_post_button') }}
+          </BooleanSetting>
+        </li>
+        <li>
+          <BooleanSetting path="autohideFloatingPostButton">
+            {{ $t('settings.autohide_floating_post_button') }}
+          </BooleanSetting>
+        </li>
+        <li>
           <BooleanSetting path="padEmoji">
             {{ $t('settings.pad_emoji') }}
           </BooleanSetting>